Refactor app theme
