--- context/darkMode.ts
@@ -1,5 +1,0 @@
-import { createContext, useContext } from 'react';
-
-export const ColorModeContext = createContext({ toggleColorMode: () => {} });
-
-export const useColorMode = () => useContext(ColorModeContext);
